Evaluate dynamically added <py-script> tag
In original version, the element created by document.createElement("py-script") after pyodide runtime initialization does not evaluate script content because tags in scriptsQueue( added on creation of the <py-script> tags) runs only on runtime initialization.
This commit changes addToScriptsQueue so that evaluate the script instantly instead of adding to the queue when it is called after runtime initialization.