Fix test_async and test_stdio_handling
Resolves #1313 and #1314. On top of #1318.

The point of these tests is to define the execution order of Tasks
that are scheduled in <py-script> tags: first all the py-script tags
are executed and their related lifecycle events. Once all of this
is done, we schedule any enqueued tasks.

To delay the execution of these tasks, we use a custom event loop for
pyExec with this defer behavior. Until `start_` is called, our custom loop
just stores the tasks. When `start_` is called, all deferred tasks are
scheduled and any later tasks are also scheduled immediately.

This deferment should only apply to user logic, so we set the custom event
loop in a context manager around the user calls and restore the original
event loop when we're done.