Import deprecated global modules lazily
There's logic to conditionally load micropip but it doesn't work
correctly (the test is always truthy so it always loads it). This
hides another problem which is that we eagerly import micropip in
order to include it in the deprecated globals. This makes the
deprecated global modules lazy imports so that we can not always
install micropip.